requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
kafka-python
pydantic
requests

python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
